requirements.txt
google-cloud-bigquery>=3.25.0
google-cloud-bigquery-datatransfer>=3.15.0
PyYAML>=6.0.0
pydantic>=2.0.0
Jinja2>=3.1.0
python-crontab>=3.0.0
